V1.1.2
Ajuste a campo identificación de cliente en el módulo de débito automático y en el servicio de débito por lo que se debe recompilar antes de desplegar

+++ Chevysat Débito Automatico BBVA/DebitoAutomatico.PS/Modulos/Configuracion/IngresarCliente.aspx.cs
@@ -465,7 +465,7 @@ namespace DebitoAutomatico.PS.Modulos.Configuracion
                                                              Convert.ToInt32(this.ddlFormaDebito.SelectedValue), true, //Formato
                                                              true, true, //Tercero
                                                              this.txbNombreTercero.Text.Trim(), //Nombre del tercero
-                                                             Convert.ToInt32(this.txbIdentificacionT.Text.Trim()), true, //Nombre del tercero
+                                                             Convert.ToInt64(this.txbIdentificacionT.Text.Trim()), true, //Nombre del tercero
                                                              Convert.ToInt32(this.ddlTipoIdentificacionT.SelectedValue), true, //Tipo ident tercero
                                                              Request.UserHostName.ToString(),
                                                              Convert.ToInt32(this.ddlFechaDebito.SelectedValue), true, //Fecha débito
